requirements.txt
flask==3.0.3
flask-socketio==5.3.6
eventlet==0.36.1
pandas==2.2.2
numpy==1.26.4
ccxt==4.4.49
python-dotenv==1.0.1
requests==2.32.3
